+++ javascript/packages/core/components/views/project/project-detail.tsx
@@ -3,7 +3,7 @@ import { useStudioQuery } from '#core/hooks/use-studio-query';
 
 export function ProjectDetail() {
   const { projectId } = useStudioParams('base');
-  const { data } = useStudioQuery<{ project: Record<string, unknown> }>({
+  const { data } = useStudioQuery<{ metadata: Record<string, string> }>({
     queryName: 'GetProject',
     serviceOptions: {
       name: projectId,
@@ -13,5 +13,5 @@ export function ProjectDetail() {
 
   // The project type will not be directly exposed to the @michelangelo/core package.
   // eslint-disable-next-line @typescript-eslint/dot-notation
-  return <div>{data?.project?.metadata?.['name']}</div>;
+  return <div>{data?.metadata?.['name']}</div>;
 }
